feat(dispd): emulador VT completo (responde DSR/DA/CPR -> destrava ash)
Reescreve vt.c como maquina de estados VT/ANSI completa (modelo st/xterm):
- responde DSR/CPR (ESC[6n) e DA (ESC[c) escrevendo de volta no pty; sem isso
  o busybox lineedit (CONFIG_FEATURE_EDITING=y) trava esperando a posicao do
  cursor antes de imprimir o prompt -> terminal preto (#111, causa-raiz do ash)
- tela primaria + alternada (?1049/47/1047), regiao de rolagem (DECSTBM),
  origin mode, autowrap adiado (DECAWM), save/restore (DECSC/DECRC, s/u)
- CSI amplo: CUU..CUP, ED/EL, IL/DL/ICH/DCH/ECH, SU/SD, VPA/CHA/CBT, TBC, REP
- SGR com 16/256/truecolor (Cell guarda COLORREF -> truecolor exato) (#114)
- decodifica UTF-8 em codepoints, 1 celula/codepoint (#106); largura CJK basica
- CSI param overflow clampeado (#108), >16 params nao sobrescreve o ultimo (#109)
- vt_init falha em OOM -> term_create retorna NULL (#105)
- rx/alive via Interlocked (#86); TERM=xterm-256color pros filhos

+++ src/dispd/compositor.c
@@ -396,7 +396,7 @@ void compose_and_present(void)
                 (g_srv.debug || w->term->rx == 0 || !w->term->alive))
                 snprintf(lbl, sizeof lbl, "%s  [%s rx:%ld a:%d]", base,
                          w->term->be ? w->term->be->name : "?",
-                         w->term->rx, w->term->alive);
+                         w->term->rx, (int)w->term->alive);
             else
                 snprintf(lbl, sizeof lbl, "%s", base);
             RECT tr = { ix + 6, iy, ix + iw - 6, iy + g_srv.title_h };
